filemanager now uses unlocker for delete retry

+++ src/core/FileManager.java
@@ -7,6 +7,7 @@ import java.util.ArrayList;
 import logger.Logger;
 import logger.ILogConfig.Module;
 
+import util.CommandTask;
 import util.FileTools;
 import util.SystemTools;
 
@@ -15,11 +16,17 @@ import util.SystemTools;
  */
 public class FileManager implements ISystemComponent {
 	
-	private Logger logger;
+	private TaskManager taskManager;
 	
-	public FileManager(Logger logger){
-		this.logger = logger;
+	private IFileConfig fileConfig;
+
+	public FileManager(TaskManager taskManager){
+		
+		this.taskManager = taskManager;
+		fileConfig = null;
 	}
+	
+	public void setConfig(IFileConfig fileConfig){ this.fileConfig = fileConfig; }
 	
 	@Override
 	public void init() throws Exception {
@@ -40,7 +47,7 @@ public class FileManager implements ISystemComponent {
 	
 	@Override
 	public void shutdown() throws Exception {
-		delete(getTempFolder());
+		FileTools.deleteFolder(getTempFolderPath());
 	}
 	
 	public String getDataFolderPath(){
@@ -75,14 +82,26 @@ public class FileManager implements ISystemComponent {
 		return new File(getLaunchFolderPath(id));
 	}
 	
-	// TODO retry with unlocker for windows
-	public void delete(File file) throws Exception {
+	public boolean hasUnlocker() {
+		return fileConfig != null && !fileConfig.getUnlocker().isEmpty();
+	}
+	
+	public void deleteWithUnlocker(File file, Logger logger) throws Exception {
 		
-		logger.debug(Module.COMMON, "delete: "+file.getAbsolutePath());
-		if(file.isFile()){
-			FileTools.deleteFile(file.getAbsolutePath());
-		}else if(file.isDirectory()){
-			FileTools.deleteFolder(file.getAbsolutePath());
+		if(hasUnlocker()){
+			String command = fileConfig.getUnlocker();
+			String arguments = file.getAbsolutePath();
+			String path = file.getParentFile().getAbsolutePath();
+			CommandTask task = new CommandTask(
+					command, arguments, path, taskManager, logger
+			);
+			task.syncRun(0, IFileConfig.UNLOCKER_TIMEOUT);
+			logger.debug(Module.COMMON, "delete: "+file.getAbsolutePath());
+			if(file.isFile()){
+				FileTools.deleteFile(file.getAbsolutePath());
+			}else if(file.isDirectory()){
+				FileTools.deleteFolder(file.getAbsolutePath());
+			}
 		}
 	}
 }
